+++ src/main/java/com/github/sadikovi/riff/tree/TypedBoundReference.java
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-package com.github.sadikovi.riff.ntree;
+package com.github.sadikovi.riff.tree;
 
 import org.apache.spark.sql.types.DataType;
