+++ src/main/java/com/github/sadikovi/riff/tree/TypedExpression.java
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-package com.github.sadikovi.riff.ntree;
+package com.github.sadikovi.riff.tree;
 
 import org.apache.spark.sql.catalyst.InternalRow;
 import org.apache.spark.sql.types.DataType;
