+++ src/test/scala/com/github/sadikovi/riff/tree/FilterSuite.scala
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-package com.github.sadikovi.riff.ntree
+package com.github.sadikovi.riff.tree
 
 import java.util.NoSuchElementException
 
@@ -30,7 +30,7 @@ import org.apache.spark.unsafe.types.UTF8String
 
 import com.github.sadikovi.riff.TypeDescription
 import com.github.sadikovi.riff.RiffTestUtils._
-import com.github.sadikovi.riff.ntree.expression._
+import com.github.sadikovi.riff.tree.expression._
 import com.github.sadikovi.testutil.UnitTestSuite
 
 class FilterSuite extends UnitTestSuite {
