+++ src/test/scala/com/github/sadikovi/riff/tree/expression/ExpressionSuite.scala
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-package com.github.sadikovi.riff.ntree.expression
+package com.github.sadikovi.riff.tree.expression
 
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.types.{IntegerType, LongType, StringType}
